Implemented asteroids splitting in two when their health hits zero.

+++ Asteroids2/Asteroids2/inizializzazioni.h
@@ -8,5 +8,5 @@ void UPDATE_VAO_Curva(Shape* fig);
 void init_player_actor(Actor* player);
 void init_projectile_actor(Actor* projectile);
 void init_background_actor(Actor* background);
-Actor* init_asteroid(float initialR, int index);
+Actor* init_asteroid(int initialR, int index);
